Build a shopping cart
Add a CART_SESSION_ID variable to the settings file that contains
a key that we are going to use to store the cart in the user session.
Create an application named cart for managing shopping carts.
Create a file named cart.py and configute a Cart class inside it
that will us to manage the shopping cart. Then create shopping cart
views to add, update and remove items from the cart. Configure urls
for these views. Create a template to display the cart.