Create an app that registrates customers' orders
Create a new app named orders. Create an Order model to store the
order details and an OrderItem model to store the items bought.
Register these models in the admin site using the ModelInline class.
Create a form for a user to enter the order details and then add
an order_create view to the views.py file that handles creation of
order objects. Configure url for this view. Create templates for
creating orders.
Install Rabbitmq and Celery to work with asynchronous tasks.
Add celery to the project and create a task that will send an email
to the user after an order has been placed.